fix: transfer the logic of contrast gamma change to the canvas-wrapper
